WWS: UserFile now only read on startup
This fixes an issue in WOOL Web Service where the user.xml file was being read for every call to a dialogue end-point... twice... Now the userfile is read on service startup and never again. So, adding new users to the user.xml file while the service is running will no longer work.